refactor: generate
